requirements.txt
fastapi
uvicorn
pytest
PyYAML
PySide6
